Extract game results xlsx into GameResultsFileInteractor
Address review: build the results workbook through an interactor instead
of composing get_full_game + get_game_stat + export_results inline in the
API route, mirroring AllGameKeysReaderInteractor.

- Move the pure table-building logic (to_results, results_to_table_routed,
  the helper types) plus a build_results_table entry point that performs
  the completeness check into shvatka/core/games/results.py. It depends
  only on core dto and the TablePrinter interface.
- Reduce infrastructure export_results to a thin wrapper over
  build_results_table + print_table so the bot keeps working unchanged.
- Add GameResultsFileInteractor(dao: GameStatReader, printer: TablePrinter)
  returning the xlsx as BinaryIO, wire it into GamePlayProvider, and use it
  from GET /games/{id}/stat/export.
